+++ modules/analysis-common/src/test/resources/org/elasticsearch/analysis/common/synonyms.json
